Add option to force username entry
Some systems (e.g. Jupyter servers) will not run code in an account whose username matches an account in EDD.  Especially for publicly-accessible notebooks where it's undesirable for security purposes to expose the username of an EDD account, it's best to force users to input a username rather than defaulting to the username of the system account used to run the code.

Also used loops in place of ‘if’ statements to force users to enter the required username/password or to tolerate accidentally pressing enter before providing it.