Update export_metadata function
Making use of the `in_study` parameter to the REST API calls, saving the
request used to convert from a slug to a Study ID. Default to using the
`active` flag on fetching Line records, to mirror behavior of the export API
endpoints (per Joonhoon request in EDD-1447). Eliminating some extraneous data
structures used during processing of the returned results.